clean code bulk import
